Restructured main folder .cpp files are now under main/src/ .hpp files are now under main/include/
Note: might be a good idea to use .hpp header files to distinguish our files from other libraries' header files
