+++ main/include/sdlog.hpp
